+++ esp_cam_sensor/sensors/sc121at/private_include/sc121at_regs.h
@@ -1,0 +1,23 @@
+/*
+ * SPDX-FileCopyrightText: 2024 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#pragma once
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#define SC121AT_REG_END      0xffff
+#define SC121AT_REG_DELAY    0xfffe
+
+/* SC121AT registers */
+#define SC121AT_REG_SENSOR_ID_H             0x3108
+#define SC121AT_REG_SENSOR_ID_L             0x3107
+
+
+#ifdef __cplusplus
+}
+#endif